RRTMGP - GPU seg fault fix
This commit is a temporary fix for the RRTMGP GPU seg faults. The
seg faults are due to data directives missing from the CAM-RRTMGP
interface. This fix will be reverted back once we add the OpenACC
directive to CAM-RRTMGP interface.

The two config options here are set to false as these check are not
needed for optimized runs.